Optimize serialization in converters
Introduce the concept of a Default CRS and Unspecified CRS as per the geojson spec i.e. if a null value is specified for a geojson crs property, this means that the CRS cannot be assumed. However, if no crs property is given then the default CRS should be used. The current implementation treats both a null CRS and the lack of a CRS as the same thing, which is incorrect. Since the library is compiled to work with Portable Class libraries, a TypeConverter cannot be used to handle initializing the GRS property on a GeoJSONObject type to a default CRS, so the serialization/deserialization handling is performed via private serialization/deserialization methods.