Don't touch CRS on (de)serializing
Removes code that creates a default CRS on deserializing when JSON contains None, and that removes an existing default CRS when serializing.

According to https://tools.ietf.org/html/rfc7946#section-4, the CRS shouldn't be specified anyway, so it's not a good idea to create one whenever deserializing a CRS-less JSON object.