Consolidate nuspec into csproj
Most information from nuspec can be specified in the csproj.
This has the benefit that all information is located in a single place and that previously duplicated information can no longer be out of sync.
Currently release notes and package dependencies are out of sync.

I've added the following:
* Xavier Fischer as Author.
* Referred MIT license directly instead of throught `LICENSE.MD`
* link to github project page
* link to git repository

csproj does not support the concepts of `<owners>` and `<language>` so I've left them out.

The following tags are automatically inferred:
* `<id>` from project name
* `requireLicenseAcceptance` defaults to `false`
* `AssemblyVersion` is `$(Version).0`

Some good resources:
* https://github.com/NuGet/Home/wiki/Adding-nuget-pack-as-a-msbuild-target
* https://andrewlock.net/version-vs-versionsuffix-vs-packageversion-what-do-they-all-mean/